Split up evaluation_template_factory
