+++ source/matrix_free/evaluation_template_factory_hanging_nodes.inst.in
@@ -17,18 +17,8 @@
 for (deal_II_dimension : DIMENSIONS;
      deal_II_scalar_vectorized : REAL_SCALARS_VECTORIZED)
   {
-    template struct dealii::internal::FEEvaluationFactory<
-      deal_II_dimension,
-      deal_II_scalar_vectorized::value_type,
-      deal_II_scalar_vectorized>;
-
-    template struct dealii::internal::FEFaceEvaluationFactory<
-      deal_II_dimension,
-      deal_II_scalar_vectorized::value_type,
-      deal_II_scalar_vectorized>;
-
-    // inverse mass
-    template struct dealii::internal::CellwiseInverseMassFactory<
+    // hanging nodes
+    template struct dealii::internal::FEEvaluationHangingNodesFactory<
       deal_II_dimension,
       deal_II_scalar_vectorized::value_type,
       deal_II_scalar_vectorized>;
